# Conflicts:
#	api-module/nuls-api-module/src/main/java/io/nuls/api/utils/DocumentTransferTool.java
#	block/nuls-block/src/main/java/io/nuls/block/message/handler/BlockHandler.java
#	block/nuls-block/src/main/java/io/nuls/block/message/handler/GetSmallBlockHandler.java
#	block/nuls-block/src/main/java/io/nuls/block/message/handler/TxGroupHandler.java
#	consensus/poc-consensus/src/main/java/io/nuls/poc/utils/manager/BlockManager.java
#	network/nuls-network/src/main/resources/module.json
#	tools/nuls-tools/api-provider/src/main/java/io/nuls/api/provider/account/AccountServiceForRpc.java
#	tools/nuls-tools/api-provider/src/main/java/io/nuls/api/provider/consensus/ConsensusProviderForRpc.java
#	tools/nuls-tools/api-provider/src/main/java/io/nuls/api/provider/transaction/TransferServiceForRpc.java
#	transaction/nuls-transaction/src/main/java/io/nuls/transaction/manager/ChainManager.java
#	transaction/nuls-transaction/src/test/java/io/nuls/transaction/TestJyc.java

+++ consensus/poc-consensus/src/main/java/io/nuls/poc/model/dto/input/CreateAgentDTO.java
@@ -11,7 +11,7 @@ public class CreateAgentDTO {
     private String agentAddress;
     private String packingAddress;
     private String rewardAddress;
-    private int commissionRate;
+    private byte commissionRate;
     private String deposit;
     private String password;
 
@@ -47,11 +47,11 @@ public class CreateAgentDTO {
         this.rewardAddress = rewardAddress;
     }
 
-    public int getCommissionRate() {
+    public byte getCommissionRate() {
         return commissionRate;
     }
 
-    public void setCommissionRate(int commissionRate) {
+    public void setCommissionRate(byte commissionRate) {
         this.commissionRate = commissionRate;
     }
